Move SsdDetector to a separate file
